server/requirements/lambda.txt
sqlalchemy >= 1.2.6
python-dotenv >= 0.7.1
marshmallow >= 3.0.4